review: extract order utils, seed scramble overlay, satisfy style guide
Addresses PR #141 review feedback:

- Extract `applyPreferredOrder` and `mergeOrders` into a shared
  `src/app/core/utils/order.utils.ts`; remove the duplicated copies
  from `my-tasks-list` and `available-tasks`.
- Seed the scramble overlay with the final text before `setText` so the
  scrambler reads a non-empty "from" state. Without this, characters
  popped in from nothing during the pre-start frames instead of
  flickering in place — a regression from the overlay refactor.
- Use `textContent = ''` rather than `innerHTML = ''` when clearing the
  overlay; avoids the HTML parser and is marginally more performant.
- Type the `onGroupDrop` return as `void` in both components.
- Wrap the drag-handle header's class list across two lines so the
  attribute stays under the 100-character style-guide limit.
