+++ src/app/features/my-tasks/components/available-tasks.component.html
@@ -34,7 +34,8 @@
         ></div>
         <header
           cdkDragHandle
-          class="flex items-center justify-between gap-2 mb-3 flex-wrap cursor-grab active:cursor-grabbing select-none"
+          class="flex items-center justify-between gap-2 mb-3 flex-wrap select-none
+                 cursor-grab active:cursor-grabbing"
           title="Drag to reorder"
         >
           <div class="flex items-center gap-2 min-w-0">
