+++ src/app/features/my-tasks/components/my-tasks-list.component.ts
@@ -10,6 +10,7 @@ import { CommonModule, DatePipe } from '@angular/common';
 import { CdkDragDrop, DragDropModule, moveItemInArray } from '@angular/cdk/drag-drop';
 
 import { Project, Task, CYBERPUNK_COLORS } from '../../../core/models/domain.model';
+import { applyPreferredOrder, mergeOrders } from '../../../core/utils/order.utils';
 
 type StatusFilter = 'all' | 'open' | 'todo' | 'in-progress' | 'done';
 
@@ -21,45 +22,6 @@ interface TaskGroup {
 }
 
 const GROUP_ORDER_STORAGE_KEY = 'omnitask:myTasks:projectGroupOrder';
-
-/**
- * Reorder `groups` so IDs that appear in `preferred` come first in that
- * relative order, followed by any groups the user hasn't yet positioned,
- * sorted by the `fallback` comparator (default: tasks-count descending).
- */
-function applyPreferredOrder<T extends { projectId: string; tasks: readonly unknown[] }>(
-  groups: T[],
-  preferred: string[],
-  fallback: (a: T, b: T) => number = (a, b) => b.tasks.length - a.tasks.length,
-): T[] {
-  if (preferred.length === 0) {
-    return groups.slice().sort(fallback);
-  }
-  const byId = new Map(groups.map((g) => [g.projectId, g] as const));
-  const ordered: T[] = [];
-  const seen = new Set<string>();
-  for (const id of preferred) {
-    const g = byId.get(id);
-    if (g) {
-      ordered.push(g);
-      seen.add(id);
-    }
-  }
-  const leftovers = groups.filter((g) => !seen.has(g.projectId)).sort(fallback);
-  return ordered.concat(leftovers);
-}
-
-/**
- * Fold the user's new visible-groups order into their stored preference
- * without forgetting IDs that aren't currently on screen (e.g. a project
- * the user filtered out). Visible IDs keep their new positions; absent
- * IDs keep their previous relative order appended at the end.
- */
-function mergeOrders(visible: string[], previous: string[]): string[] {
-  const visibleSet = new Set(visible);
-  const tail = previous.filter((id) => !visibleSet.has(id));
-  return visible.concat(tail);
-}
 
 /**
  * Compact, cross-project list of the user's tasks. Grouped by project so the
@@ -158,7 +120,7 @@ export class MyTasksListComponent {
    * in their new order — unseen projects stay where their old preference
    * (if any) put them so the user's memory of the layout stays stable.
    */
-  onGroupDrop(event: CdkDragDrop<TaskGroup[]>) {
+  onGroupDrop(event: CdkDragDrop<TaskGroup[]>): void {
     if (event.previousIndex === event.currentIndex) return;
     const groups = [...this.groupedTasks()];
     moveItemInArray(groups, event.previousIndex, event.currentIndex);
